requirements_1.txt
python-dotenv~=1.0.1
pip install google-generativeai
